Add hog feature extraction
HOG is a simple and powerful feature descriptor. It is not only
used for face detection but also it is widely used for object
detection like cars, pets, and fruits. HOG is robust for object
detection because object shape is characterized using the local
intensity gradient distribution and edge direction.